Trait futhark::ArrayBase

pub trait ArrayBase {
    type Raw;
    type Ctx: Context;
#[must_use = "Don't forget to check the error code"]    unsafe fn free(
        ctx: *mut <Self::Ctx as Context>::Raw, 
        raw: *mut Self::Raw
    ) -> c_int;
}

Basic information about a Futhark array, and functions for manipulating it.

Associated Types

type Raw

The C type used to represent arrays with this type.

type Ctx: Context

The futhark_context type arrays with this ArrayBase can be used with.

Loading content...
Required methods

#[must_use = "Don't forget to check the error code"]unsafe fn free(
ctx: *mut <Self::Ctx as Context>::Raw,
raw: *mut Self::Raw
) -> c_int

Free the underlying array.

Safety

The provided raw pointer must point to a valid array.

Loading content...