Trait futhark::Context

pub trait Context {
    type Raw;
    fn raw(&self) -> *mut Self::Raw;

    unsafe fn last_error(&self) -> *const c_char;

    unsafe fn sync(&self) -> c_int;

    fn check_error(&self, code: c_int) -> Result<(), FutharkError> { ... }

    fn null_check<T>(&self, ptr: *mut T) -> Result<(), FutharkError> { ... }
}

A wrapper struct containing a futhark_context.

Associated Types

type Raw

The futhark_context type used by the generated bindings.

Loading content...
Required methods

fn raw(&self) -> *mut Self::Raw

Get a pointer to the underlying context type.

unsafe fn last_error(&self) -> *const c_char

unsafe fn sync(&self) -> c_int

Loading content...
Provided methods

fn check_error(&self, code: c_int) -> Result<(), FutharkError>

A helper method for converting an error code into a FutharkError, using Context::last_error() to extract the error message.

fn null_check<T>(&self, ptr: *mut T) -> Result<(), FutharkError>

A helper method for converting an error code into a FutharkError, using Context::last_error() to extract the error message.

Loading content...