Trait futhark::ExtractableArray

pub trait ExtractableArray: ArrayBase {
    type Shape: AsRef<[usize]> + AsMut<[usize]> + Default;
    type Item: Copy;
    unsafe fn shape(
        ctx: *mut <Self::Ctx as Context>::Raw, 
        raw: *mut Self::Raw
    ) -> *const i64;

#[must_use = "Don't forget to check the error code"]    unsafe fn to_vec(
        ctx: *mut <Self::Ctx as Context>::Raw, 
        raw: *mut Self::Raw, 
        buffer: *mut Self::Item
    ) -> c_int;
}

An array that lets you access the items inside it.

Associated Types

type Shape: AsRef<[usize]> + AsMut<[usize]> + Default

The type used to represent an array's shape (e.g. [usize; 2] for a 2D array).

type Item: Copy

The item a buffer would contain.

Loading content...
Required methods

unsafe fn shape(
ctx: *mut <Self::Ctx as Context>::Raw,
raw: *mut Self::Raw
) -> *const i64

Get a reference to the array containing this array's shape.

Safety

This must return a pointer to an i64 array with the same length as ExtractableArray::Shape.

#[must_use = "Don't forget to check the error code"]unsafe fn to_vec(
ctx: *mut <Self::Ctx as Context>::Raw,
raw: *mut Self::Raw,
buffer: *mut Self::Item
) -> c_int

Copy the contents of this array to a buffer.

Safety

The buffer must have space for the entire array's contents. This can be calculated as the product of the ExtractableArray::Shape lengths.

Loading content...